osutil/disks: allow building on mac os
This is to allow to continue compiling the snap command on darwin, as very soon
the snap command imports bootloader, and the bootloader package will start
importing the disks package. None of the functionality from the snap command
however depends on actually using the disks package so it's sufficient to just
provide stub implementations here.

Note that to actually build the snap command on darwin, one must define the tags
nomanagers and nosecboot with -tags nosecboot,nomanagers.
